Cleaning EntryArgs
